refactor context usage
