--- src/renderer/context/editor/index.ts
@@ -1,7 +1,0 @@
-/**
- * Index file for editor-related contexts
- */
-export { default as EditorProvider } from './EditorContext';
-export { useEditor } from './EditorContext';
-export { default as AltoEditorProvider } from './AltoEditorContext';
-export { useAltoEditor } from './AltoEditorContext';
