+++ src/renderer/context/index.ts
@@ -1,19 +1,5 @@
-/**
- * Main index file for all context providers and hooks
- */
-
-// Project contexts
-export { ProjectProvider, useProject } from './project';
-
-// Editor contexts
-export { 
-  EditorProvider, 
-  useEditor, 
-  AltoEditorProvider, 
-  useAltoEditor,
-} from './editor';
-
-// App contexts
-export { SettingsProvider, useSettings, AltoProvider, useAlto } from './app';
-
-// Note: The old AltoContext has been refactored and moved to the app directory 
+export * from './AltoContext';
+export * from './AltoEditorContext';
+export * from './EditorContext';
+export * from './ProjectContext';
+export * from './SettingsContext';
