--- src/renderer/context/project/index.ts
@@ -1,5 +1,0 @@
-/**
- * Index file for project-related contexts
- */
-export { default as ProjectProvider } from './ProjectContext';
-export { useProject } from './ProjectContext';
